Do not use MailMerge
Today, I made the fatal mistake of trusting Microsoft Office to do anything right. Not sure what possessed me to do that, but it won't happen again. I have written a script that sends emails to everyone who won items, automatically deciding whether to send a receipt or a reminder. It can also now send emails to multiple people and refuses to send emails to the empty-string. Best of all, it actually sends emails, instead of putting them in one's outbox and then immediately closing Outlook so that they can't send.
I also cleaned up the LaTeX code; it was leaving files open and making use of overly-verbose syntax.